client/requirements.txt
APScheduler==2.1.2
CherryPy==3.2.2
PocketSphinx==0.8
PyYAML==3.10
Pykka==1.2.0
RPi.GPIO==0.5.4
SphinxBase==0.8
argparse==1.2.1
beautifulsoup4==4.2.1
facebook-sdk==0.4.0
feedparser==5.1.3
numpy==1.8.0
pifacecommon==4.0.0
pifacedigitalio==3.0.4
pygame==1.9.1release
python-dateutil==2.1
python-mpd==0.3.0
pytz==2013b
quantities==0.10.1
semantic==1.0.1
six==1.6.1
ws4py==0.3.2
wsgiref==0.1.2
